# agent-api/openapi_gateway.yml
openapi: "3.0.3"
info:
  title: Agente Simple
  description: Agente conversacional con LangChain create_agent, FastAPI y PostgreSQL.
  version: "1.0.0"

# Extensión de Google API Gateway: declara los backends con nombre.
# Referencia: https://docs.cloud.google.com/api-gateway/docs/oasv3-extensions
x-google-api-management:
  backends:
    agent_api_backend:
      address: https://agent-api-service-846286405720.us-central1.run.app
      pathTranslation: APPEND_PATH_TO_ADDRESS
      protocol: "http/1.1"

x-google-backend: agent_api_backend

servers:
  - url: https://agent-gateway-ass12pug.uc.gateway.dev
    x-google-endpoint:
      allowCors: True

tags:
  - name: chat
    description: Operaciones de conversación con el agente
  - name: history
    description: Historial de conversaciones y sesiones
  - name: errors
    description: Catálogo de tipos de error
  - name: feedback
    description: Calificación de respuestas del agente

paths:
  /api/chat:
    options:
      operationId: chatOptions
      summary: CORS preflight
      x-google-backend: agent_api_backend
      responses:
        "204":
          description: No Content
    post:
      summary: Enviar un mensaje al agente
      description: |
        Recibe una pregunta del usuario y retorna la respuesta del agente.
        Si `session_id` no se envía o viene vacío, se genera un UUID aleatorio.
        El historial de conversación se recupera desde la base de datos usando el `session_id`.
      operationId: chat
      tags:
        - chat
      x-google-backend: agent_api_backend
      security:
        - firebase: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            examples:
              nueva_sesion:
                summary: Nueva sesión (sin session_id)
                value:
                  question: "¿Cuál es la fecha de hoy?"
                  user: "igorov"
              sesion_existente:
                summary: Sesión existente
                value:
                  question: "¿Y qué hora es?"
                  user: "igorov"
                  session_id: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Respuesta del agente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
              example:
                user: igorov
                answer: "Hoy es 25 de mayo de 2026."
                session_id: "550e8400-e29b-41d4-a716-446655440000"
                trace_id: "123e4567-e89b-12d3-a456-426614174000"
        "401":
          description: Token de autenticación ausente o inválido.
        "422":
          description: Error de validación en el request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "500":
          description: Error interno del servidor

  /api/history/{session_id}:
    parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
        description: ID de sesión cuyo historial se desea consultar
        example: "550e8400-e29b-41d4-a716-446655440000"
    options:
      operationId: getHistoryOptions
      summary: CORS preflight
      x-google-backend: agent_api_backend
      responses:
        "204":
          description: No Content
    get:
      summary: Obtener historial de conversación
      description: Retorna todos los mensajes de una sesión ordenados cronológicamente.
      operationId: getHistory
      tags:
        - history
      x-google-backend: agent_api_backend
      security:
        - firebase: []
      responses:
        "200":
          description: Lista de mensajes del historial
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryItem"
        "401":
          description: Token de autenticación ausente o inválido.
        "422":
          description: Error de validación en el request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "500":
          description: Error interno del servidor

  /api/sessions/{user}:
    parameters:
      - name: user
        in: path
        required: true
        schema:
          type: string
        description: Identificador del usuario
        example: "igorov"
    options:
      operationId: getSessionsByUserOptions
      summary: CORS preflight
      x-google-backend: agent_api_backend
      responses:
        "204":
          description: No Content
    get:
      summary: Obtener sesiones de un usuario
      description: Retorna los session_id únicos de un usuario ordenados del más reciente al más antiguo.
      operationId: getSessionsByUser
      tags:
        - history
      x-google-backend: agent_api_backend
      security:
        - firebase: []
      responses:
        "200":
          description: Lista de session_id del usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSessionsResponse"
              example:
                user: "igorov"
                sessions:
                  - "550e8400-e29b-41d4-a716-446655440000"
                  - "123e4567-e89b-12d3-a456-426614174001"
        "401":
          description: Token de autenticación ausente o inválido.
        "422":
          description: Error de validación en el request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "500":
          description: Error interno del servidor

  /api/errors:
    options:
      operationId: listErrorTypesOptions
      summary: CORS preflight
      x-google-backend: agent_api_backend
      responses:
        "204":
          description: No Content
    get:
      summary: Listar tipos de error
      description: Retorna el catálogo de tipos de error disponibles para calificar una respuesta incorrecta.
      operationId: listErrorTypes
      tags:
        - errors
      x-google-backend: agent_api_backend
      security:
        - firebase: []
      responses:
        "200":
          description: Lista de tipos de error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorTypeResponse"
        "401":
          description: Token de autenticación ausente o inválido.
        "500":
          description: Error interno del servidor

  /api/feedback:
    options:
      operationId: feedbackOptions
      summary: CORS preflight
      x-google-backend: agent_api_backend
      responses:
        "204":
          description: No Content
    post:
      summary: Calificar una respuesta del agente
      description: |
        Recibe la calificación del usuario sobre una respuesta del agente identificada por `trace_id`.
        Si `is_ok` es `false`, se puede indicar el tipo de error (`error_type_id`) y observaciones adicionales.
      operationId: feedback
      tags:
        - feedback
      x-google-backend: agent_api_backend
      security:
        - firebase: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
            examples:
              calificacion_ok:
                summary: Respuesta correcta
                value:
                  trace_id: "123e4567-e89b-12d3-a456-426614174000"
                  is_ok: true
              calificacion_error:
                summary: Respuesta incorrecta con tipo de error
                value:
                  trace_id: "123e4567-e89b-12d3-a456-426614174000"
                  is_ok: false
                  error_type_id: 1
                  observations: "La respuesta no fue correcta para mi caso."
      responses:
        "200":
          description: Feedback guardado correctamente, retorna el registro actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "401":
          description: Token de autenticación ausente o inválido.
        "404":
          description: trace_id no encontrado
        "422":
          description: Error de validación en el request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "500":
          description: Error interno del servidor

components:
  # Esquemas de seguridad. Para Firebase Auth se utiliza el flujo OAuth2 implícito
  # junto con la extensión x-google-auth (issuer, jwksUri, audiences) propia de
  # OpenAPI v3 en Google API Gateway.
  # Ref: https://docs.cloud.google.com/api-gateway/docs/oasv3-extensions
  securitySchemes:
    firebase:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://pragmatic-ratio-478805-u2.firebaseapp.com
          scopes: {}
      x-google-auth:
        issuer: https://securetoken.google.com/pragmatic-ratio-478805-u2
        jwksUri: https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com
        audiences:
          - pragmatic-ratio-478805-u2

  schemas:
    ChatRequest:
      type: object
      required:
        - question
        - user
      properties:
        question:
          type: string
          description: Pregunta o mensaje del usuario
          example: "¿Cuál es la fecha de hoy?"
        user:
          type: string
          description: Identificador del usuario
          example: igorov
        session_id:
          type: string
          nullable: true
          description: ID de sesión para continuar una conversación. Si está vacío o ausente, se genera uno aleatorio.
          example: "550e8400-e29b-41d4-a716-446655440000"

    ChatResponse:
      type: object
      required:
        - user
        - answer
        - session_id
        - trace_id
      properties:
        user:
          type: string
          description: Identificador del usuario
          example: igorov
        answer:
          type: string
          description: Respuesta generada por el agente
          example: "Hoy es 25 de mayo de 2026."
        session_id:
          type: string
          description: ID de sesión (generado si no se envió)
          example: "550e8400-e29b-41d4-a716-446655440000"
        trace_id:
          type: string
          description: ID único de trazabilidad generado por petición
          example: "123e4567-e89b-12d3-a456-426614174000"

    HistoryItem:
      type: object
      required:
        - question
        - answer
        - trace_id
        - session_id
        - created_at
      properties:
        question:
          type: string
          description: Pregunta del usuario
        answer:
          type: string
          description: Respuesta del agente
        trace_id:
          type: string
          description: ID único de trazabilidad
          example: "123e4567-e89b-12d3-a456-426614174000"
        session_id:
          type: string
          description: ID de sesión
          example: "550e8400-e29b-41d4-a716-446655440000"
        user:
          type: string
          nullable: true
          description: Identificador del usuario
          example: "igorov"
        retrieved_contexts:
          type: array
          nullable: true
          description: Contextos recuperados usados por el agente para responder
          items:
            type: object
            properties:
              text:
                type: string
              source:
                type: string
                nullable: true
              score:
                type: number
                nullable: true
              chunk_id:
                type: string
                nullable: true
        created_at:
          type: string
          format: date-time
          description: Fecha y hora de creación del registro
        is_ok:
          type: boolean
          nullable: true
          description: Indica si la respuesta fue calificada como correcta
        error_type_id:
          type: integer
          nullable: true
          description: ID del tipo de error registrado en el feedback
        observations:
          type: string
          nullable: true
          description: Observaciones adicionales del feedback

    ErrorTypeResponse:
      type: object
      required:
        - id
        - description
      properties:
        id:
          type: integer
          description: Identificador del tipo de error
          example: 1
        description:
          type: string
          description: Descripción del tipo de error
          example: "Respuesta incorrecta"

    FeedbackRequest:
      type: object
      required:
        - trace_id
        - is_ok
      properties:
        trace_id:
          type: string
          description: ID de trazabilidad de la respuesta a calificar
          example: "123e4567-e89b-12d3-a456-426614174000"
        is_ok:
          type: boolean
          description: Indica si la respuesta fue correcta
          example: false
        error_type_id:
          type: integer
          nullable: true
          description: ID del tipo de error (requerido si is_ok es false)
          example: 1
        observations:
          type: string
          nullable: true
          description: Descripción adicional del error
          example: "La respuesta no fue relevante para mi consulta."

    FeedbackResponse:
      type: object
      required:
        - trace_id
        - session_id
        - question
        - answer
        - created_at
      properties:
        trace_id:
          type: string
        session_id:
          type: string
        question:
          type: string
        answer:
          type: string
        user:
          type: string
          nullable: true
          description: Identificador del usuario
          example: "igorov"
        retrieved_contexts:
          type: array
          nullable: true
          description: Contextos recuperados usados por el agente para responder
          items:
            type: object
            properties:
              text:
                type: string
              source:
                type: string
                nullable: true
              score:
                type: number
                nullable: true
              chunk_id:
                type: string
                nullable: true
        created_at:
          type: string
          format: date-time
        is_ok:
          type: boolean
          nullable: true
        error_type_id:
          type: integer
          nullable: true
        observations:
          type: string
          nullable: true

    UserSessionsResponse:
      type: object
      required:
        - user
        - sessions
      properties:
        user:
          type: string
          description: Identificador del usuario
          example: "igorov"
        sessions:
          type: array
          description: Lista de session_id ordenados del más reciente al más antiguo
          items:
            type: string
          example:
            - "550e8400-e29b-41d4-a716-446655440000"
            - "123e4567-e89b-12d3-a456-426614174001"

    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string